+++ sevenbridges/errors.py
@@ -43,6 +43,14 @@ class ValidationError(SbgError):
         )
 
 
+class TaskValidationError(SbgError):
+    def __init__(self, message, task=None):
+        self.task = task
+        super(TaskValidationError, self).__init__(
+            code=-1, status=-1, message=message
+        )
+
+
 class PaginationError(SbgError):
     def __init__(self, message):
         super(PaginationError, self).__init__(
@@ -88,6 +96,20 @@ class TooManyRequests(SbgError):
     def __init__(self, code=None, message=None, more_info=None):
         super(TooManyRequests, self).__init__(
             code=code, status=429, message=message, more_info=more_info
+        )
+
+
+class ServerError(SbgError):
+    def __init__(self, code=None, message=None, more_info=None):
+        super(ServerError, self).__init__(
+            code=code, status=500, message=message, more_info=more_info
+        )
+
+
+class ServiceUnavailable(SbgError):
+    def __init__(self, code=None, message=None, more_info=None):
+        super(ServiceUnavailable, self).__init__(
+            code=code, status=503, message=message, more_info=more_info
         )