+++ CHANGES.md
@@ -1,3 +1,26 @@
+0.5.1 (2016-08-29)
+==================
+- Bugfixes
+  - [Issue #28](https://github.com/sbg/sevenbridges-python/issues/28)
+  
+0.5.0 (2016-08-29)
+==================
+
+- Project settings available on project resource. These settings
+  can be currently used to create locked projects.
+- File now have tags. User is able to view/set/edit tags.
+- Support for proxies. User can now configure library to use proxies.
+
+
+- Bugfixes
+  - [Issue #30](https://github.com/sbg/sevenbridges-python/issues/30)
+  - [Issue #29](https://github.com/sbg/sevenbridges-python/issues/29)
+  - [Issue #28](https://github.com/sbg/sevenbridges-python/issues/28)
+  - [Issue #24](https://github.com/sbg/sevenbridges-python/issues/24)
+  - [Issue #23](https://github.com/sbg/sevenbridges-python/issues/23)
+  - [Issue #14](https://github.com/sbg/sevenbridges-python/issues/14)
+
+
 0.4.4 (2016-08-19)
 ==================
 - Upload changes